backend/requirements.txt
# backend/requirements.txt
# Backend Dependencies for Pokweb

# Web Framework
fastapi==0.110.0
uvicorn==0.22.0

# Database
sqlalchemy==1.4.46
mysql-connector-python==8.0.33
alembic==1.10.3
pydantic==2.6.3
pydantic-settings==2.2.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Email
aiosmtplib==2.0.1
jinja2==3.1.2

# Utilities
python-multipart==0.0.6
email-validator

# Security
cryptography==40.0.2

# Optional but recommended
python-dotenv==1.0.0
